no anda nada
el codigo esta tan roto que no se puede avanzar nada.. voy a hacer un
commit con las ultimas cosas y voy a reconstruir todo con codigo viejo e
ir avanzando a medida que anden cosas. es la unica manera

+++ Source/fuentes/configurador.c
@@ -24,7 +24,7 @@ void iniciar_sysclock (void)
 }
 
 
-void seleccionar_puerto(int puerto)
+void seleccionar_puerto(unsigned short puerto)
 {
 	switch (puerto)
 	{
